Add doubles

+++ apps/scala/src/ppl/apps/liszt/gputest/MatFieldTest.scala
@@ -6,16 +6,16 @@ import ppl.dsl.deliszt.MetaInteger._
 object MatFieldTestRunner extends DeLisztApplicationRunner with MatFieldTest
 
 trait MatFieldTest extends DeLisztApplication {
-  var int3_zero  : Rep[Vec[_3,Int]] = null
-  var int33_zero : Rep[Mat[_3,_3,Int]] = null
-  var PMat       : Rep[Field[Edge,Mat[_3,_3,Int]]] = null
-  var edge_values: Rep[Field[Edge,Mat[_3,_3,Int]]] = null
+  var int3_zero  : Rep[Vec[_3,Double]] = null
+  var int33_zero : Rep[Mat[_3,_3,Double]] = null
+  var PMat       : Rep[Field[Edge,Mat[_3,_3,Double]]] = null
+  var edge_values: Rep[Field[Edge,Mat[_3,_3,Double]]] = null
     
   def main() {
-    int3_zero  = Vec(1,2,3)
+    int3_zero  = Vec(1.0,2.0,3.0)
     int33_zero = Mat(int3_zero,int3_zero,int3_zero)
-    PMat = FieldWithConst[Edge,Mat[_3,_3,Int]](int33_zero)
-    edge_values = FieldWithConst[Edge,Mat[_3,_3,Int]](int33_zero)
+    PMat = FieldWithConst[Edge,Mat[_3,_3,Double]](int33_zero)
+    edge_values = FieldWithConst[Edge,Mat[_3,_3,Double]](int33_zero)
     
     for(e <- edges(mesh)) {
       edge_values(e) += (PMat(e) + int33_zero)
